+++ programs/uxd/src/instructions/credix_lp/edit_credix_lp_depository.rs
@@ -48,7 +48,7 @@ pub struct EditCredixLpDepositoryFields {
     pub redeeming_fee_in_bps: Option<u8>,
     pub minting_disabled: Option<bool>,
     pub profits_beneficiary_collateral: Option<Pubkey>,
-    pub redeemable_amount_under_management_weight: Option<u64>,
+    pub redeemable_amount_under_management_weight: Option<u32>,
 }
 
 pub(crate) fn handler(
@@ -151,6 +151,7 @@ pub(crate) fn handler(
             redeemable_amount_under_management_weight
         });
     }
+
     Ok(())
 }
